requirements.txt
asgiref==3.2.4
Django==3.0.5
django-debug-toolbar==2.2
django-environ==0.4.5
django-extensions==2.2.8
django-import-export==2.1.0
djangorestframework==3.11.0
ipython==7.13.0
mysqlclient==1.4.6
Pillow==7.0.0
pytz==2019.3
sqlparse==0.3.1
werkzeug==1.0.0
